math: Fix UB in ldbl-128 llrintl
testing _Float128 (without inline functions)
UBSAN: Undefined behaviour in ../sysdeps/ieee754/float128/../ldbl-128/s_llrintl.c:83:31 left shift of 281474976710656 by 15 cannot be represented in type 'long long int'

+++ sysdeps/ieee754/ldbl-128/s_llrintl.c
@@ -80,7 +80,7 @@ __llrintl (_Float128 x)
       else if (j0 <= 48)
 	result = i0 >> (48 - j0);
       else
-	result = ((long long int) i0 << (j0 - 48)) | (i1 >> (112 - j0));
+	result = (i0 << (j0 - 48)) | (i1 >> (112 - j0));
     }
   else
     {
